Fix unnecessary `-extensions v3_ca`. Compat with OpenSSL 3.3. Fixes #1.
`-extensions v3_ca` was supported (and ingnored) by the `openssl req`
of OpenSSL < 3.3 only due to a bug, see:

* https://github.com/openssl/openssl/issues/22966
* https://github.com/openssl/openssl/pull/16865

See how Python fixed it similarly:

* https://github.com/python/cpython/commit/744caa8ef42ab67c6aa20cd691e078721e72e22a

In our code, the options that are problematic for `openssl req`
but need to be set for `openssl x509` are set explicitly with
the inline-generated `.ini` files.

It still works with OpenSSL 3.0.